separate add and remove - add tests for both
